feat: move link handler types to a shared location
