+++ apps/web/src/modules/dao/components/About/About.tsx
@@ -3,6 +3,7 @@ import { useDaoMembership } from '@buildeross/hooks/useDaoMembership'
 import { getFetchableUrls } from '@buildeross/ipfs-service'
 import { metadataAbi, tokenAbi } from '@buildeross/sdk/contract'
 import { SubgraphSDK } from '@buildeross/sdk/subgraph'
+import { type ProfileLinkHandler } from '@buildeross/types'
 import { Avatar } from '@buildeross/ui/Avatar'
 import { FallbackImage } from '@buildeross/ui/FallbackImage'
 import { unpackOptionalArray } from '@buildeross/utils/helpers'
@@ -17,7 +18,7 @@ import useSWR from 'swr'
 import { Address, formatEther } from 'viem'
 import { useAccount, useBalance, useReadContracts } from 'wagmi'
 
-import { MembersList, type ProfileLinkHandler } from '../MembersList'
+import { MembersList } from '../MembersList'
 import { responsiveGrid } from './About.css'
 import { DaoDescription } from './DaoDescription'
 import { ExternalLinks } from './ExternalLinks'
