+++ apps/web/src/modules/dao/components/Activity/Activity.tsx
@@ -5,7 +5,7 @@ import { useDelegate } from '@buildeross/hooks/useDelegate'
 import { useQueryParams } from '@buildeross/hooks/useQueryParams'
 import { useVotes } from '@buildeross/hooks/useVotes'
 import { getProposals, ProposalsResponse } from '@buildeross/sdk/subgraph'
-import { CHAIN_ID } from '@buildeross/types'
+import { CHAIN_ID, ProposalLinkHandler } from '@buildeross/types'
 import { Countdown } from '@buildeross/ui/Countdown'
 import { AnimatedModal, SuccessModalContent } from '@buildeross/ui/Modal'
 import { walletSnippet } from '@buildeross/utils/helpers'
@@ -15,7 +15,7 @@ import React from 'react'
 import { ContractButton } from 'src/components/ContractButton'
 import Pagination from 'src/components/Pagination'
 import { Upgrade, useProposalStore } from 'src/modules/create-proposal'
-import { ProposalCard, ProposalLinkHandler } from 'src/modules/proposal'
+import { ProposalCard } from 'src/modules/proposal'
 import { useChainStore, useDaoStore } from 'src/stores'
 import { skeletonAnimation } from 'src/styles/animations.css'
 import { sectionWrapperStyle } from 'src/styles/dao.css'
