+++ apps/web/src/modules/dao/components/DaoCard/DaoCard.tsx
@@ -2,23 +2,17 @@ import { PUBLIC_DEFAULT_CHAINS } from '@buildeross/constants/chains'
 import { useCountdown } from '@buildeross/hooks/useCountdown'
 import { useIsMounted } from '@buildeross/hooks/useIsMounted'
 import { getFetchableUrls } from '@buildeross/ipfs-service'
-import { AddressType, CHAIN_ID } from '@buildeross/types'
+import { AddressType, CHAIN_ID, DaoLinkHandler } from '@buildeross/types'
 import { BigNumberish, formatCryptoVal } from '@buildeross/utils/numbers'
 import { Box, Flex, Paragraph, Text } from '@buildeross/zord'
 import dayjs from 'dayjs'
 import NextImage from 'next/image'
 import React, { useState } from 'react'
 import { FallbackNextImage } from 'src/components/FallbackNextImage'
-import { LinkWrapper as Link, LinkWrapperOptions } from 'src/components/LinkWrapper'
+import { LinkWrapper as Link } from 'src/components/LinkWrapper'
 
 import { auction, daoImage, name, title } from './DaoCard.css'
 import { Detail } from './Detail'
-
-export type DaoLinkHandler = (
-  chainId: CHAIN_ID,
-  tokenAddress: AddressType,
-  tokenId?: number | string | bigint
-) => LinkWrapperOptions
 
 interface DaoCardProps {
   chainId: CHAIN_ID
