+++ apps/web/src/pages/dao/[network]/[token]/index.tsx
@@ -59,7 +59,7 @@ const DaoPage: NextPageWithLayout<DaoPageProps> = ({ chainId, collectionAddress 
   )
 
   const getProposalLink = React.useCallback(
-    (proposalNumber: number) => {
+    (_chainId: CHAIN_ID, _collectionAddress: AddressType, proposalNumber: number) => {
       return {
         href: `/dao/${chain.slug}/${addresses.token}/vote/${proposalNumber}`,
       }
