+++ packages/types/src/index.ts
@@ -5,3 +5,4 @@ export * from './duration'
 export * from './hex'
 export * from './migration'
 export * from './proposalState'
+export * from './links'
